Reduce code complexity in Analytics settings view.
- remove eslint ignore comment
- extract logic from SettingsView into individual components.
